Documenting the structure of the data responses. (#7012)
Documenting the structure of the /web/api/data responses (Fixes #7005)

Swagger2.0 semi-supports this - they understand the sub-typing but do not show
the sub-types in the model of the response directly. There are comments in
each of the cases to guide the user towards the correct part of the model. For each 
kind of format request we have a data model of the response.

Miscellaneous style improvements and typo fixes.